подключает модули
