+++ index.html
@@ -235,5 +235,9 @@
     </template>
     <script src="./js/main.js">
     </script>
+    <script type="module" src="./js/data.js">
+    </script>
+     <script type="module" src="./js/util.js">
+    </script>
 </body>
 </html>
